fix: refuse a hosted image url that is not a public https url
`run` fetches every distinct hosted image URL before dispatching anything,
to check the picture is still the one that was evaluated. It fetched
whatever the envelope said, with no check on the URL itself, so an
envelope naming `http://10.0.0.5/shot.png` or `http://169.254.169.254/`
produced a plaintext request to that address and nothing reported it.

`redirectDowngradeReason` does not cover this. Its rule is that a request
must not end up somewhere less safe than it started, and there is no
downgrade from a URL that was never https. Admitting a URL and following
one are separate questions.

`validateImageEnvelope` now requires `url` to be a public https URL when
present, so the envelope is refused at load and no request is issued.
That is the same rule `post()` already applies when it stores the URL;
this applies it again on the way back in, because the envelope is a
hand-editable file read long after it was written. It is the reasoning
`asBareFilename` beside it already uses for the sibling `file` field.

Behaviour change worth naming: an envelope hand-edited to point at a
local Shutterbug's image is now refused on load. `post()` already refuses
to create one, so no workflow depends on it.

`isPublicHttpsUrl` moves to `src/urls.ts` because `schemas.ts` needs it
and `backends/types.ts`, where it lived, imports `schemas.ts`. The
functions are unchanged; `backends/types.ts` re-exports them, so
`shutterbug.ts`, `execute.ts` and their tests are untouched.

Reported in review of the milestone-2 follow-ups PR. The fix there was to
check inside `hostedImageCheck`, which would have failed every test in
`hosted-image-check.test.ts` -- all four drive a real loopback server to
exercise streaming, size limits and hashing against a real socket.

Six rejection cases and two acceptance cases; all six fail with the
call site reverted.
